requirements.txt
Django>=4.2,<6.1
Pillow>=10.0,<11.0
whitenoise>=6.0
gunicorn>=21.0
python-dotenv>=1.0.0
django-allauth>=65.0
requests>=2.31.0
xhtml2pdf>=0.2.12
djangorestframework>=3.14.0